Start work on a non-generic, column based representation.
Instead of using templatized types for all of the tasks, sources
and sinks, this uses a fixed "column group" representation that
supports common primitive types and strings.

This should be flexible enough to support most use cases efficiently;
since we operate on groups of values at a time, we can perform
operations like sorting, grouping and combining without having to
perform large numbers of virtual method calls or pointer deferences.